requirements.txt
# requirements.txt — Breast Cancer Detection Pipeline
# Install with: pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
flask>=3.0.0
flask-cors
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
Pillow>=10.0.0
albumentations>=1.3.0
grad-cam>=1.5.0
seaborn>=0.13.0
lime
shap
python-docx
